requirements.txt
Pillow==10.2.0
pyautogui==0.9.53
scikit-image==0.22.0
